Recover panic in Block unmarshal functions
There are no checks for data length in unmarshal functions, so it used to just panic when data was smaller than
expected. It is also impossible to check the length at once, since the proper length is calculated dynamically
depending on data itself, and with multiple checks it seems to slow down too much, so we go for recover anti-pattern.