Features: reset voting list correctly
We must not delete anything from the database as a result of *adding blocks*, because this operation is irreversible.
Irreversible operations can not be rollbacked (we remove the data, there is no way to restore what was in DB).
Therefore, if rollback covered voting period edge, feature votes would not be restored at all.
Instead, we should add new records with 0 votes at the end of voting period.