Accounts Data Storage: make immediate entries updates
When Data tx is performed, it is expected that entries appear in state immediately, since
they can be used by following transactions even in same block.
In old implementation entries are written to DB batch, appearing in state only after this batch
is written.
In new implementation entries are written to local storage (and returned from there) and at the end
of applying blocks batch are written to DB batch.