Exchange tx: only use pointers for Orders (for all the versions)
Before that, both pointer and value types for Order implemented Order interface.
But when casting in OrderToOrderBody and other functions, only value was accepted.
This leads to errors when creating exchange with pointer, which is also valid Order.
The good solution is to fixate that only pointer implements Order interface.